ufs: fix compilation warnings

+++ drivers/scsi/ufs/ufs-exynos.c
@@ -1129,7 +1129,7 @@ static int exynos_ufs_link_startup_notify(struct ufs_hba *hba, bool notify)
 {
 	int ret = 0;
 
-	switch (notify) {
+	switch ((u8)notify) {
 	case PRE_CHANGE:
 		exynos_ufs_dev_hw_reset(hba);
 		ret = exynos_ufs_pre_link(hba);
@@ -1150,7 +1150,7 @@ static int exynos_ufs_pwr_change_notify(struct ufs_hba *hba, bool notify,
 {
 	int ret = 0;
 
-	switch (notify) {
+	switch ((u8)notify) {
 	case PRE_CHANGE:
 		ret = exynos_ufs_pre_prep_pmc(hba, pwr_max, pwr_req);
 		break;
@@ -1167,7 +1167,7 @@ static int exynos_ufs_pwr_change_notify(struct ufs_hba *hba, bool notify,
 static void exynos_ufs_hibern8_notify(struct ufs_hba *hba,
 				u8 enter, bool notify)
 {
-	switch (notify) {
+	switch ((u8)notify) {
 	case PRE_CHANGE:
 		exynos_ufs_pre_hibern8(hba, enter);
 		break;
